Break nw_content.* into separate files
* browser/nw_content_browser_hooks.*: hooks only depending on content API for browser process
* renderer/nw_content_renderer_hooks.*: hooks only depending on content API for renderer process
* browser/nw_extensions_browser_hooks.*: hooks depending on extensions and content API for browser process
* renderer/nw_extensions_renderer_hooks.*: hooks depending on extensions and content API for renderer process
* browser/nw_chrome_browser_hooks.*: hooks depending on chrome, extensions and content API for browser process
* renderer/nw_chrome_renderer_hooks.*: hooks depending on chrome, extensions and content API for renderer process
* common/nw_content_common_hooks.*: hooks depending on content, extensions and content API for browser or renderer process
* common/node_hooks.*: hooks helper for Node.js
