+++ nw.gypi
@@ -117,7 +117,20 @@
         'src/api/shell/shell.h',
         'src/api/tray/tray.cc',
         'src/api/tray/tray.h',
-        'src/nw_content.cc',
+        'src/browser/nw_chrome_browser_hooks.cc',
+        'src/browser/nw_chrome_browser_hooks.h',
+        'src/browser/nw_content_browser_hooks.cc',
+        'src/browser/nw_content_browser_hooks.h',
+        'src/browser/nw_extensions_browser_hooks.cc',
+        'src/browser/nw_extensions_browser_hooks.h',
+        'src/common/node_hooks.h',
+        'src/common/nw_content_common_hooks.cc',
+        'src/common/nw_content_common_hooks.h',
+        'src/renderer/nw_chrome_renderer_hooks.h',
+        'src/renderer/nw_content_renderer_hooks.cc',
+        'src/renderer/nw_content_renderer_hooks.h',
+        'src/renderer/nw_extensions_renderer_hooks.cc',
+        'src/renderer/nw_extensions_renderer_hooks.h',
         'src/nw_content.h',
         'src/nw_custom_bindings.cc',
         'src/nw_custom_bindings.h',
