+++ src/renderer/nw_chrome_renderer_hooks.h
@@ -1,0 +1,35 @@
+#ifndef NW_CHROME_RENDERER_HOOKS_H_
+#define NW_CHROME_RENDERER_HOOKS_H_
+
+#include "third_party/WebKit/public/web/WebNavigationPolicy.h"
+
+namespace blink {
+  class WebFrame;
+  class WebString;
+  class WebURLRequest;
+}
+
+namespace content {
+  class RenderView;
+}
+
+namespace extensions {
+  class Dispatcher;
+}
+
+namespace nw {
+//
+// implemented in nw_extensions_renderer_hooks.cc
+//
+// ref in chrome/renderer/chrome_content_renderer_client.cc
+void willHandleNavigationPolicy(content::RenderView* rv,
+                               blink::WebFrame* frame,
+                               const blink::WebURLRequest& request,
+                               blink::WebNavigationPolicy* policy,
+                               blink::WebString* manifest,
+                               bool new_win);
+// ref in chrome/renderer/extensions/chrome_extensions_renderer_client.cc
+void ExtensionDispatcherCreated(extensions::Dispatcher* dispatcher);
+}
+
+#endif // NW_CHROME_RENDERER_HOOKS_H_
